requirements.txt
Flask
Flask-PyMongo
Flask-Swagger-UI
pytest
python-dotenv
mongomock